@@ -3,30 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.21.1",
     "bootstrap": "^5.0.1",
-    "react": "^17.0.2",
+    "react": "^19.2.0",
     "react-bootstrap": "^1.6.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^19.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
     "uuid": "^8.3.2",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest",
+    "test:ui": "vitest --ui"
   },
   "browserslist": {
     "production": [
@@ -40,7 +33,12 @@
       "last 1 safari version"
     ]
   },
-  "jest": {
-    "resetMocks": false
+  "devDependencies": {
+    "@testing-library/jest-dom": "^6.9.1",
+    "@vitejs/plugin-react": "^5.0.4",
+    "@vitest/ui": "^3.2.4",
+    "jsdom": "^27.0.0",
+    "vite": "^7.1.8",
+    "vitest": "^3.2.4"
   }
 }
